Wave A: lint our register headers + tests/mocks; strip 2.3k redundant casts
Drop ra8d2_*_regs and tests/mocks from .clang-tidy HeaderFilterRegex --
both are first-party code and should be linted. The previous exclusion
silenced ~2300 redundant-casting findings from typed-enum constants,
e.g. (uint16_t)k_some_enum where the enum is already `: uint16_t`.

clang-tidy --fix removed all of them across 244 files. Builds and the
full host-test suite are unchanged; the casts were genuinely no-ops.
Cleanup of the auto-fix collateral:

- The fixer used C++ `static_cast<int>` syntax in C files when widening
  bool to int, then redundantly wrapped expressions like `if (X)` as
  `(int)(X)`. Removed the wrappers entirely (C accepts any scalar in
  conditional context) and dropped the redundant parens (which v22's
  readability-redundant-parentheses correctly flagged).
- `readability-implicit-bool-conversion` flagged idiomatic C `return x
  != y;` because relational operators in C produce `int` per the C
  spec, and its v18 auto-fixer wrapped them in `(x != y) != 0` --
  identical noise. Reverted the 99 spurious wrappers and disabled the
  check with a comment explaining the C/C++ semantic mismatch.

Three computed-cast sites in ra_rsip.c / ra8d2_mipi_csi_regs.h needed
NOLINT for clang-analyzer-optin.core.EnumCastOutOfRange (NOLINT count
42 -> 57) -- the analyzer can't see that a base + stride lands on a
HUM-defined register, only that the literal value isn't an enumerator.

Per-line rationale comments now sit above each disabled check in
.clang-tidy so the next reader knows why each one is off.

Tidy now passes on both clang-tidy 18 (Ubuntu/CI/docker) and 22
(Homebrew/macOS); 86/86 host tests pass; blink + blink_hal sizes
unchanged.

+++ blink/main.c
@@ -183,12 +183,12 @@ static void ra_systick_stop(void)
 static void ra_systick_wait_ticks(uint32_t ticks)
 {
   ra_systick_stop();
-  *ra_reg32(k_ra_systick_rvr) = (ticks - 1U) & (uint32_t)k_ra_systick_rvr_max;
+  *ra_reg32(k_ra_systick_rvr) = (ticks - 1U) & k_ra_systick_rvr_max;
   *ra_reg32(k_ra_systick_cvr) = 0U;
-  *ra_reg32(k_ra_systick_csr) = (uint32_t)k_ra_systick_enable | (uint32_t)k_ra_systick_clksource;
+  *ra_reg32(k_ra_systick_csr) = k_ra_systick_enable | k_ra_systick_clksource;
 
   /* Poll COUNTFLAG. Goes 1 on wrap (one-shot), auto-clears on read. */
-  while ((*ra_reg32(k_ra_systick_csr) & (uint32_t)k_ra_systick_countflag) == 0U) {
+  while ((*ra_reg32(k_ra_systick_csr) & k_ra_systick_countflag) == 0U) {
     /* wait */
   }
 
@@ -212,10 +212,9 @@ static void ra_systick_wait_ticks(uint32_t ticks)
  */
 static void ra_delay_ms(uint32_t ms)
 {
-  uint32_t total = (ms * (uint32_t)k_ra_cpu_hz_at_reset) / (uint32_t)k_ra_blink_ms_per_sec;
+  uint32_t total = (ms * k_ra_cpu_hz_at_reset) / k_ra_blink_ms_per_sec;
   while (total > 0U) {
-    const uint32_t chunk =
-      (total > (uint32_t)k_ra_systick_rvr_max) ? (uint32_t)k_ra_systick_rvr_max : total;
+    const uint32_t chunk = (total > k_ra_systick_rvr_max) ? k_ra_systick_rvr_max : total;
     ra_systick_wait_ticks(chunk);
     total -= chunk;
   }
@@ -237,14 +236,14 @@ static void ra_blink_pin_init(uint8_t port, uint8_t bit)
   volatile uint8_t* pwprs = ra_reg8(k_ra_pmisc_pwprs);
 
   *pwpr  = 0x00U;
-  *pwpr  = (uint8_t)k_ra_pwpr_pfswe;
+  *pwpr  = k_ra_pwpr_pfswe;
   *pwprs = 0x00U;
-  *pwprs = (uint8_t)k_ra_pwpr_pfswe;
+  *pwprs = k_ra_pwpr_pfswe;
 
-  *ra_pfs_reg(port, bit) = (uint32_t)k_ra_pfs_pdr_output_low;
+  *ra_pfs_reg(port, bit) = k_ra_pfs_pdr_output_low;
 
-  *pwpr  = (uint8_t)k_ra_pwpr_b0wi;
-  *pwprs = (uint8_t)k_ra_pwpr_b0wi;
+  *pwpr  = k_ra_pwpr_b0wi;
+  *pwprs = k_ra_pwpr_b0wi;
 }
 
 /**
@@ -261,7 +260,7 @@ static void ra_blink_drive_all(uint8_t high)
   uint32_t pdr_mask[k_ra_blink_pin_count]   = {};
   uint32_t port_count                       = 0U;
 
-  for (uint32_t i = 0U; i < (uint32_t)k_ra_blink_pin_count; ++i) {
+  for (uint32_t i = 0U; i < k_ra_blink_pin_count; ++i) {
     const uint8_t port = k_ra_blink_pins[i].port;
     const uint8_t bit  = k_ra_blink_pins[i].bit;
 
@@ -296,7 +295,7 @@ static void ra_blink_drive_all(uint8_t high)
 #pragma GCC diagnostic ignored "-Wmain"
 int32_t main(void)
 {
-  for (uint32_t i = 0U; i < (uint32_t)k_ra_blink_pin_count; ++i) {
+  for (uint32_t i = 0U; i < k_ra_blink_pin_count; ++i) {
     ra_blink_pin_init(k_ra_blink_pins[i].port, k_ra_blink_pins[i].bit);
   }
 
@@ -306,9 +305,9 @@ int32_t main(void)
 
   while (1) {
     ra_blink_drive_all(1U);
-    ra_delay_ms((uint32_t)k_ra_blink_half_period_ms);
+    ra_delay_ms(k_ra_blink_half_period_ms);
     ra_blink_drive_all(0U);
-    ra_delay_ms((uint32_t)k_ra_blink_half_period_ms);
+    ra_delay_ms(k_ra_blink_half_period_ms);
   }
 
   return 0;
