+++ libs/ra_hal/src/ra_cac.c
@@ -87,8 +87,7 @@ ra_err_t ra_cac_measure(uint16_t* out_count)
  * @brief combined mask.
  */
 typedef enum : uint8_t {
-  k_ra_cac_status_mask_all =
-    (uint8_t)k_ra_cac_status_mendf | (uint8_t)k_ra_cac_status_ovff | (uint8_t)k_ra_cac_status_ferrf,
+  k_ra_cac_status_mask_all = k_ra_cac_status_mendf | k_ra_cac_status_ovff | k_ra_cac_status_ferrf,
 } ra_cac_bits_w43_t;
 
 /**
@@ -117,7 +116,7 @@ ra_err_t ra_cac_deinit(void)
 ra_err_t ra_cac_get_status(uint8_t* out_mask)
 {
   RA_CHECK_NULL_PTR(out_mask, s_tag, "out_mask must not be nullptr");
-  *out_mask = (uint8_t)(ra_cac()->CASTR & (uint8_t)k_ra_cac_status_mask_all);
+  *out_mask = (uint8_t)(ra_cac()->CASTR & k_ra_cac_status_mask_all);
   return k_ra_ok;
 }
 
@@ -126,7 +125,7 @@ ra_err_t ra_cac_clear_status(uint8_t mask)
   volatile r_cac_regs_t* reg = ra_cac();
   /* HUM Ch 26.2.4 CAICR bit positions (FERRFCL/MENDFCL/OVFFCL) are
    * offset by 4 from CASTR bits -- write 1 to clear. */
-  reg->CAICR = (uint8_t)(mask & (uint8_t)k_ra_cac_status_mask_all);
+  reg->CAICR = (uint8_t)(mask & k_ra_cac_status_mask_all);
   return k_ra_ok;
 }
 
@@ -140,7 +139,7 @@ ra_err_t ra_cac_attach_handler(ra_cac_event_fn_t fn, void* ctx)
 void ra_cac_dispatch(void)
 {
   volatile r_cac_regs_t* reg  = ra_cac();
-  const uint8_t          mask = (uint8_t)(reg->CASTR & (uint8_t)k_ra_cac_status_mask_all);
+  const uint8_t          mask = (uint8_t)(reg->CASTR & k_ra_cac_status_mask_all);
   reg->CAICR                  = mask;
   const ra_cac_event_fn_t fn  = s_cac_state.fn;
   void* const             ctx = s_cac_state.ctx;
